renaming home and discovery features modules
