+++ discovery-feature/build.gradle.kts
@@ -1,6 +1,9 @@
 import extensions.applyAndroidDefault
+import extensions.arrow
 import extensions.commonBaseDependencies
+import extensions.exoPlayer
 import extensions.featureBaseDependencies
+import extensions.implementationProject
 
 plugins {
     id(GradlePlugins.androidLibrary)
@@ -14,4 +17,17 @@ applyAndroidDefault()
 dependencies {
     commonBaseDependencies()
     featureBaseDependencies()
+
+    implementation(Libraries.viewPager)
+    implementation(Libraries.coilImageLoader)
+
+    exoPlayer()
+
+    // Arrow
+    arrow()
+
+    implementationProject(
+        Modules.dataSources,
+        Modules.entities
+    )
 }
