+++ discovery-feature/src/main/res/drawable/background_game_player.xml
