+++ discovery-feature/src/main/res/drawable/background_game_player_overlay.xml
