+++ discovery-feature/src/main/res/layout/view_carousel_player.xml
