+++ discovery-feature/src/main/res/values/colors.xml
