+++ discovery-feature/src/main/res/values/dimens.xml
