+++ home-feature/build.gradle.kts
@@ -1,9 +1,6 @@
 import extensions.applyAndroidDefault
-import extensions.arrow
 import extensions.commonBaseDependencies
-import extensions.exoPlayer
 import extensions.featureBaseDependencies
-import extensions.implementationProject
 
 plugins {
     id(GradlePlugins.androidLibrary)
@@ -17,17 +14,4 @@ applyAndroidDefault()
 dependencies {
     commonBaseDependencies()
     featureBaseDependencies()
-
-    implementation(Libraries.viewPager)
-    implementation(Libraries.coilImageLoader)
-
-    exoPlayer()
-
-    // Arrow
-    arrow()
-
-    implementationProject(
-        Modules.dataSources,
-        Modules.entities
-    )
 }
