+++ home-feature/lint-baseline.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<issues format="5" by="lint 4.0.0" client="gradle" variant="debug" version="4.0.0">
+<issues format="5" by="lint 4.0.1" client="gradle" variant="debug" version="4.0.1">
 
 </issues>
