+++ home-feature/src/main/AndroidManifest.xml
